# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "",
        "version": "... BOOKPERFECT API DESIGNED BY HOTELISTAN COMPANY ..."
    },
    "servers": [
        {
            "url": "https://cms.hitrooms.com/rest-api"
        }
    ],
    "paths": {
...
        "/get__travelidea_ideaid_lan_currency/{idea}": {
            "get": {
                "tags": [
                    "@Ideas and Packages"
                ],
                "summary": "RETURN TRAVELIDEA BY IDEA_ID AND LANGUAGE AND CURRENCY",
                "operationId": "41",
                "parameters": [
                    {
                        "name": "idea",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
        "/get__TripIdeasCategories__by__name/{city}": {
            "get": {
                "tags": [
                    "@TripIdeasCategories"
                ],
                "summary": "GET TRIPIDEAS BY CITY NAME",
                "operationId": "57",
                "parameters": [
                    {
                        "name": "city",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    }
                }
            }
        },
...
    }
}